fix: copy tnpm files (#1390)
* fix: fix path problem when using npm components in custom component

* fix: fix usingComponent config when npm component refers another npm component

* chore: tidy function code

* refactor: tidy code in jsx2mp-loader
