--- packages/jsx2mp-loader/src/moduleResolve.js
@@ -1,99 +1,0 @@
-const {
-  sep,
-  join,
-  dirname
-} = require('path');
-const {
-  existsSync,
-  statSync,
-  readFileSync
-} = require('fs-extra');
-
-function startsWith(prevString, nextString) {
-  return prevString.indexOf(nextString) === 0;
-}
-function removeExtension(filePath) {
-  const index = filePath.lastIndexOf('.');
-  return filePath.substr(0, index);
-}
-
-function loadAsFile(module, extensions) {
-  if (existsSync(module) && statSync(module).isFile()) {
-    return removeExtension(module);
-  }
-  for (let e of extensions) {
-    if (existsSync(module + e) && statSync(module + e).isFile()) {
-      return module;
-    }
-  }
-}
-
-function loadAsDirectory(module, extensions) {
-  if (!existsSync(module)) {
-    return;
-  }
-  let stat = statSync(module);
-  if (stat.isDirectory()) {
-    for (let e of extensions) {
-      if (existsSync(module + `/index${e}`) && statSync(module + `/index${e}`).isFile()) {
-        return join(module, '/index');
-      }
-    }
-  } else if (stat.isFile()) {
-    return loadAsFile(module, extensions);
-  }
-}
-
-function nodeModulesPaths(start) {
-  let parts = start.split(sep);
-
-  if (!parts[parts.length - 1]) {
-    parts.pop();
-  }
-
-  let i = parts.length - 1;
-  let dirs = [];
-  while (i >= 0) {
-    if ('node_modules' === parts[i]) {
-      i -= 1;
-      continue;
-    }
-    let dir = join(parts.slice(0, i + 1).join(sep) || sep, 'node_modules');
-    dirs.push(dir);
-    i -= 1;
-  }
-  return dirs;
-}
-
-function loadNpmModules(module, start, extension) {
-  let target;
-  let paths = nodeModulesPaths(start);
-
-  for (let i = 0; i < paths.length; ++i) {
-    let dependencyPath = join(paths[i], module);
-    target = loadAsFile(dependencyPath, extension) || loadAsDirectory(dependencyPath, extension);
-
-    if (target) {
-      break;
-    }
-  }
-  return target;
-}
-
-/**
-* Resolve node path.
-* @param script
-* @param dependency
-* @param extension
-* @return {*}
-*/
-module.exports = function resolve(script, dependency, extensions = ['.js', '.jsx']) {
-  let target;
-  if (startsWith(dependency, './') || startsWith(dependency, '/') || startsWith(dependency, '../')) {
-    let dependencyPath = join(script, dependency);
-    target = loadAsFile(dependencyPath, extensions) || loadAsDirectory(dependencyPath, extensions);
-  } else {
-    target = loadNpmModules(dependency, dirname(script), extensions);
-  }
-  return target;
-};
