+++ packages/jsx2mp-loader/src/utils/pathHelper.js
